requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
psutil==5.9.5
python-dotenv==1.0.0
schedule==1.2.0 